+++ android/app/src/main/java/com/backpackingmap/backpackingmap/map/ZoomLevel.kt
@@ -1,6 +1,6 @@
 package com.backpackingmap.backpackingmap.map
 
-data class ZoomLevel(val metersPerPixel: Double) {
+data class ZoomLevel(val metersPerPixel: Float) {
     fun scaledBy(factor: Float) = ZoomLevel(metersPerPixel * factor)
 }
